Split keepalive packets into separate files
