+++ src/protocol/ClientBus.ts
@@ -1,11 +1,9 @@
 import { Socket, Server } from "net";
 import { Constants } from "../Configuration";
-import { Console } from "../game/Console";
 import { Player } from "../game/Player";
 import { Client, ClientState } from "./Client";
-import { ClientboundPacket } from "./Packet";
 import { ReadableBuffer } from "./ReadableBuffer";
-import { ServerKeepAlivePacket } from "./states/play/KeepAlivePacket";
+import { ServerKeepAlivePacket } from "./states/play/ServerKeepAlivePacket";
 import { PlayerInfoActions, PlayerInfoPacket } from "./states/play/PlayerInfoPacket";
 
 export class ClientBus {
